requirements.txt
Pillow==11.2.1
pillow-heif==0.22.0
